requirements.txt
accelerate==0.32.1
affine @ file:///home/conda/feedstock_root/build_artifacts/affine_1733762038348/work
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiohttp-retry==2.8.3
aioice==0.10.1
aiortc==1.9.0
aiosignal==1.3.2
aiosqlite==0.21.0
albucore==0.0.24
albumentations==2.0.8
annotated-types @ file:///home/conda/feedstock_root/build_artifacts/annotated-types_1733247046149/work
anthropic==0.49.0
anyio==4.9.0
APScheduler==3.11.0
async-timeout==5.0.1
asyncua==1.1.6
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1741918516150/work
av==12.3.0
backoff==2.2.1
bitsandbytes==0.43.3
black==25.1.0
boto3==1.35.60
botocore==1.35.99
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-plugins @ file:///home/conda/feedstock_root/build_artifacts/click-plugins_1733731077999/work
cligj @ file:///home/conda/feedstock_root/build_artifacts/cligj_1733749956636/work
coloredlogs==15.0.1
commonmark==0.9.1
contourpy==1.3.2
cryptography==45.0.4
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1733332471406/work
Cython==3.0.12
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
einops==0.8.0
eval_type_backport @ file:///home/conda/feedstock_root/build_artifacts/eval_type_backport_1734857727836/work
exceptiongroup==1.3.0
fastapi==0.110.3
filelock==3.13.1
flake8==7.3.0
flatbuffers==25.2.10
fonttools @ file:///home/conda/feedstock_root/build_artifacts/fonttools_1651017726246/work
frozenlist==1.7.0
fsspec==2024.6.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.175.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-crc32c==1.7.1
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.1.3
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.32.6
humanfriendly==10.0
idna==3.10
ifaddr==0.2.0
imagecodecs @ file:///croot/imagecodecs_1734436728732/work
imageio @ file:///home/conda/feedstock_root/build_artifacts/imageio_1738273805233/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
iniconfig==2.1.0
Jinja2==3.1.4
jiter==0.10.0
jmespath==1.0.1
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1748019130050/work
kagglehub==0.3.12
kiwisolver @ file:///home/conda/feedstock_root/build_artifacts/kiwisolver_1648854394530/work
lazy_loader @ file:///home/conda/feedstock_root/build_artifacts/lazy-loader_1733636780672/work
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib @ file:///croot/matplotlib-suite_1713336378214/work
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.4
munkres==1.1.4
mypy_extensions==1.1.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.13.0.11
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.1.105
onnxruntime==1.19.2
opencv-contrib-python==4.10.0.84
opencv-python-headless==4.11.0.86
packaging==24.2
paho-mqtt==1.6.1
pandas==2.2.3
pathspec==0.12.1
peft==0.11.1
pillow @ file:///croot/pillow_1738010226202/work
platformdirs==4.3.8
pluggy==1.6.0
prometheus-fastapi-instrumentator==6.0.0
prometheus_client==0.22.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1653089181607/work
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.0.2
pybboxes==0.2.0
pycocotools==2.0.6
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.2.1
pydantic_core==2.33.2
pydot==2.0.0
pyee==13.0.0
pyflakes==3.4.0
Pygments==2.19.1
PyJWT==2.10.1
pylibsrtp==0.12.0
pylogix==1.0.5
pymodbus==3.8.3
pyOpenSSL==25.1.0
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1652235407899/work
PyQt5==5.15.4
PyQt5-sip==12.9.0
pytest==8.4.0
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
python-dotenv==1.0.1
pytz==2025.2
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1648757090803/work
rasterio @ file:///croot/rasterio_1740069178893/work
redis==5.0.8
regex==2024.11.6
requests==2.32.4
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.0.1
rsa==4.9.1
s3transfer==0.10.4
safetensors==0.5.3
scikit-image==0.24.0
scikit-learn @ file:///home/conda/feedstock_root/build_artifacts/scikit-learn_1652976692230/work
scipy @ file:///croot/scipy_1747238027288/work/dist/scipy-1.15.3-cp310-cp310-linux_x86_64.whl#sha256=2a791554880ad4f358fcc4cd2a982ffe1e9d472e9241011216b2be797457f1f9
shapely==2.0.7
shellingham==1.5.4
simsimd==6.4.9
sip @ file:///home/conda/feedstock_root/build_artifacts/sip_1646101165171/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
slack_sdk==3.33.5
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.37.2
stringzilla==3.12.5
structlog==24.4.0
sympy==1.13.3
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1741878222898/work
tifffile @ file:///home/conda/feedstock_root/build_artifacts/tifffile_1734094708987/work
timm==1.0.15
tldextract==5.1.3
tokenizers==0.21.1
toml @ file:///home/conda/feedstock_root/build_artifacts/toml_1734091811753/work
tomli @ file:///home/conda/feedstock_root/build_artifacts/tomli_1733256695513/work
torch==2.4.0
torchaudio==2.7.1+cu128
torchvision==0.19.0
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1648827254365/work
tqdm @ file:///croot/tqdm_1738943501192/work
transformers==4.52.4
triton==3.0.0
twilio==9.3.8
typer==0.12.5
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
ultralytics==8.3.152
ultralytics-thop==2.0.14
unicodedata2 @ file:///home/conda/feedstock_root/build_artifacts/unicodedata2_1649111910890/work
uritemplate==4.2.0
urllib3==2.4.0
wait_for2==0.3.2
yarl==1.20.1
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1748277385522/work
zxing-cpp==2.2.0
